requirements.txt
asgiref==3.7.2
boto3==1.28.17
botocore==1.31.17
coverage==7.2.7
Django==4.2.1
django-filter==23.2
django-storages==1.13.2
jmespath==1.0.1
Pillow==9.5.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
